ntlm: Move the low level winbind functions out of the HTTP winbind module
Note: All three functions are currently accessed from the http winbind
module, however the initialisation and response generation functions
will be removed in a future commit only to be called from within their
own source file.

As such we cannot define these two functions as static. as they should
be but given that this is only temporary, we have not named them to
match our naming convension as they will only need to be renamed back.
